Added function to WeatherViz
